Partial implementation of WebcastRoomPinMessage
Will work for only WebcastChatMessage pinned messages right now.
Will provide a debug of the pinned type for anything it can't decode that is pinned.
Will send a live "ChatEvent" for now with the message prefixed "<pinned>" that was pinned. Later iterations need to return a new kind of Pinned event type.