Fully removing polling and reconnection code.
Reconnection code didn't work well with pings and retry's giving akward errors, and also the websocket reference wasn't async safe when reconecting, and it seemed like ping routines could possibly leak.

It is more practical for reconnect and polling logic to be in control of the library user in any case.  May make this a more robust opt in
feature in the future. Possibly adding a is live poller that client
libraries can use then just call the track method again when appropriate.