Add GetLiveRoomUserInfo and IsLive method.
IsLive requires a LiveRoomUserInfo (may have different ways later) to quickly determine if a user is live.  This should
be used instead of simply calling a new TrackUser/Room every time to save on requests to tiktok. Reuse a
LiveRoomUserInfo for the same user.

Note, I think roomID remains the same after the very first time a user goes live, but I am not 100% sure. Please report
any issues and be specific as possible.