Instead of event stream being a generic interface type it is now of the new Event interface type.
All events are now guaranteed to have a created timestamp.  For accuracy and backwards compatability library clients
should use "CreatedTimestamp()"

The main purpose of this is library clients can keep track of the most recent created timestamp for various events and
do self filtering in case of disconnects and reconnecting by creating a new live using the track methods.